Website Framework
A website designed to host meetup details for groups interested in My Little Pony. This was a small project which I worked on to develop an understanding of HTML, CSS, SQL and how to seamlessly integrate data stored in a database into a website using Flask. As a result, the HTML and CSS is not written with accessibility in mind or documented to a high standard.

See the READ ME for more details.